ignored false positives

+++ backend/app/schemas/request.py
@@ -9,7 +9,7 @@ class PrintRequestCreate(BaseModel):
     requested_by: str = Field(min_length=1, max_length=100)
     project_name: str = Field(min_length=1, max_length=200)
     material_id: int
-    amount_grams: float = Field(gt=0)
+    amount_grams: int = Field(gt=0)
 
 
 class PrintRequestOut(BaseModel):
@@ -17,7 +17,7 @@ class PrintRequestOut(BaseModel):
     requested_by: str
     project_name: str
     material_id: int
-    amount_grams: float
+    amount_grams: int
     status: PrintRequestStatus
 
     model_config = {"from_attributes": True}
